Print OpenCL build log warnings and fix narrowing conversion
- Add build log retrieval after clBuildProgram to display actual
  warning text from the OpenCL compiler (previously only showed
  "1 warning generated" without details)
- Fix atomic_inc return type: uint -> uint (was implicitly narrowed
  to uchar, likely the warning source)
- Fix stale comment (W[64] -> rolling W[16])

+++ kernel/kernel_profanity.hpp
@@ -499,7 +499,7 @@ __kernel void profanity_iterate_score(
 
 		if(scorePrefix >= prefixCount && scoreSuffix >= suffixCount){
 			uchar score = scoreMax + 1;
-			uchar hasResult = atomic_inc(&pResult[score].found);
+			uint hasResult = atomic_inc(&pResult[score].found);
 			if (hasResult == 0) {
 				pResult[score].foundId = id;
 				for (int i = 0; i < 20; ++i) {
